+++ mobile/src/app/auth/version-check.tsx
@@ -10,7 +10,7 @@ import {useDeeplink} from "@/contexts/DeeplinkContext"
 import {useAuth} from "@/contexts/AuthContext"
 import {useAppTheme} from "@/utils/useAppTheme"
 import {loadSetting, saveSetting, SETTINGS_KEYS} from "@/utils/SettingsHelper"
-import bridge from "@/bridge/Bridge"
+import bridge from "@/bridge/MantleBridge"
 import {translate} from "@/i18n"
 import {TextStyle, ViewStyle} from "react-native"
 import {ThemedStyle} from "@/theme"
