+++ mobile/src/app/pairing/glasseswifisetup/scan.tsx
@@ -2,7 +2,7 @@ import React, {useState, useEffect, useRef} from "react"
 import {View, Text, FlatList, TouchableOpacity, ActivityIndicator, BackHandler} from "react-native"
 import {useLocalSearchParams, router, useFocusEffect} from "expo-router"
 import {Screen, Header, Button, Icon} from "@/components/ignite"
-import bridge from "@/bridge/Bridge"
+import bridge from "@/bridge/MantleBridge"
 import GlobalEventEmitter from "@/utils/GlobalEventEmitter"
 import {useAppTheme} from "@/utils/useAppTheme"
 import {ThemedStyle} from "@/theme"
