+++ mobile/src/app/pairing/loading.tsx
@@ -3,7 +3,7 @@ import {View, Text, TouchableOpacity, ScrollView, ViewStyle, TextStyle, BackHand
 import {useRoute} from "@react-navigation/native"
 import Icon from "react-native-vector-icons/FontAwesome"
 import {useCoreStatus} from "@/contexts/CoreStatusProvider"
-import coreCommunicator, {CoreCommunicator} from "@/bridge/Bridge"
+import coreCommunicator, {CoreCommunicator} from "@/bridge/MantleBridge"
 import PairingDeviceInfo from "@/components/misc/PairingDeviceInfo"
 import GlassesTroubleshootingModal from "@/components/misc/GlassesTroubleshootingModal"
 import GlassesPairingLoader from "@/components/misc/GlassesPairingLoader"
@@ -16,6 +16,7 @@ import {Header} from "@/components/ignite/Header"
 import {PillButton} from "@/components/ignite/PillButton"
 import {useNavigationHistory} from "@/contexts/NavigationHistoryContext"
 import GlobalEventEmitter from "@/utils/GlobalEventEmitter"
+import bridge from "@/bridge/MantleBridge"
 
 export default function GlassesPairingGuideScreen() {
   const {replace, clearHistory} = useNavigationHistory()
@@ -61,7 +62,7 @@ export default function GlassesPairingGuideScreen() {
   }, [handleForgetGlasses])
 
   const handlePairFailure = (error: string) => {
-    bridge.getInstance().sendForgetSmartGlasses()
+    bridge.sendForgetSmartGlasses()
     replace("/pairing/failure", {error: error, glassesModelName: glassesModelName})
   }
