+++ mobile/src/app/pairing/prep.tsx
@@ -10,7 +10,7 @@ import {Button, Header} from "@/components/ignite"
 import {router} from "expo-router"
 import {useAppTheme} from "@/utils/useAppTheme"
 import {Screen} from "@/components/ignite/Screen"
-import bridge from "@/bridge/Bridge"
+import bridge from "@/bridge/MantleBridge"
 import {translate} from "@/i18n"
 import {useNavigationHistory} from "@/contexts/NavigationHistoryContext"
 import {LinearGradient} from "expo-linear-gradient"
