+++ mobile/src/app/settings/buffer-debug.tsx
@@ -2,7 +2,7 @@ import React, {useState} from "react"
 import {View} from "react-native"
 import {Screen, Header, Text} from "@/components/ignite"
 import {useAppTheme} from "@/utils/useAppTheme"
-import bridge from "@/bridge/Bridge"
+import bridge from "@/bridge/MantleBridge"
 import ActionButton from "@/components/ui/ActionButton"
 import {Spacer} from "@/components/misc/Spacer"
 import {useNavigationHistory} from "@/contexts/NavigationHistoryContext"
